Tree inverting without building a proper tree
Since LeetCode gives a flat vector as input and expects a flat vector as
output, we don't really need to bother with a real tree structure.
Instead, since we're flipping all the nodes at a given depth, we just
need to call `reverse` on a coll of each "layer." So we reduce across
the input list and make sure we're doubling the size of each previous
step.

I did mess around with a "proper" solution using `clojure.zip` and the
node struct provided by LeetCode's sample code. I'll
commit and comment on that separately.