Add solution for middle of linked list problem
I know that traditionally this would be done with some kind of
pointer-y looping solution where you keep track of two things as you go
through the list. But this seemed more Clojure-ish to me. It's easy
enough to do the other way, but I'm enjoying another chance to use
`iterate` :)