Solve middle of linked list with pair of pointers
This is the traditional way of doing it, and actually, it is indeed a
lot more elegant. I wanted to do it the other way first because it was
what came to mind first. This was from a comment on Leetcode, which I
only saw because I was confused by their test output being [3 4 5].

I think it's important to solve it yourself rather than copying, but
since I'm writing out the test cases first, I needed to understand that.