Inefficient way of deleting items from linked list
Currently creating a vector in the filter loop, then converting that to
a linked list. So we're effectively iterating through the list twice,
when the real answer should only require one iteration. With immutable
nodes, I should be solving the problem recursively, but I wanted to
start with a straightforward `loop` first, then compare the recursive
solution to that.