Add old non-working WIP code
Going to use this for some live coding with the CIDER debugger. Maybe
we'll get it working!

I'm pretty sure the problems all stem from the difference between `nil`
and `(->TreeNode nil)`. Since I initially wrote the TreeNode stuff for a
balanced binary tree, a node either had zero children, two, or was the
sole node with one child for an uneven number of leaves. But
intentionally creating unbalanced trees, as this problem requires, means
you need a clear "this space intentionally left blank" marker. I've
haphazardly changed some things to get closer to that, but I think it
will be much simpler with a better design that accounts for this.